Added Digital Sensors

+++ src/NeutronUVSensor.cpp
@@ -6,9 +6,9 @@ NeutronUVSensor::NeutronUVSensor(String name): NeutronSensor(0, name) {
 };
 
 String NeutronUVSensor::readSensor() {
-    float visible = sensor->readVisible();
-    float ir = sensor->readIR();
-    float UVIndex = sensor->readUV() / 100;
+    int visible = sensor->readVisible();
+    int ir = sensor->readIR();
+    int UVIndex = sensor->readUV() / 100;
 
     return String(visible) + "," + String(ir) + "," + String(UVIndex);
